requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.0
cycler==0.12.1
filelock==3.19.1
flatbuffers==25.12.19
fonttools==4.60.2
fastapi==0.115.6
fsspec==2025.10.0
h11==0.16.0
humanfriendly==10.0
idna==3.11
importlib_resources==6.5.2
Jinja2==3.1.6
kiwisolver==1.4.7
MarkupSafe==3.0.3
matplotlib==3.9.4
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.2.1
numpy==2.0.2
onnx==1.19.1
onnxruntime-gpu==1.19.2
onnxslim==0.1.82
opencv-python==4.13.0.90
packaging==26.0
pillow==11.3.0
polars==1.36.1
polars-runtime-32==1.36.1
protobuf==6.33.4
psutil==7.2.1
pyparsing==3.3.2
pyreadline3==3.5.4
python-dotenv==1.0.1
python-multipart==0.0.20
pymysql==1.1.1
python-dateutil==2.9.0.post0
PyYAML==6.0.3
requests==2.32.5
scipy==1.13.1
six==1.17.0
sqlalchemy==2.0.36
sympy==1.13.1
tensorrt @ file:///C:/Program%20Files/TensorRT-10.11.0.33/python/tensorrt-10.11.0.33-cp39-none-win_amd64.whl#sha256=2983dd45c2610b2968a18e7a6aa8b8f51008bacc2d74bb2dfdd4a4c0e37a799a
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
typing_extensions==4.15.0
ultralytics==8.4.8
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.39.0
zipp==3.23.0
# WebRTC for live_detection_feed (optional: real-time video)
aiortc==1.6.0
av==14.0.0
